refactor: Concert, User, Review 제외 연관관계 해

+++ src/main/java/com/example/book_your_seat/seat/domain/Zone.java
@@ -20,7 +20,7 @@ public enum Zone {
         return expression.apply(price);
     }
 
-    public static Zone setZone(int seatNumber) {
+    public static Zone setZone(Long seatNumber) {
         if (seatNumber <= 30) {
             return SPECIAL;
         }
